add code about member, modify code of post and page

+++ baseball2/src/main/java/yc/team8/baseball/post/controller/PostController.java
@@ -1,11 +1,16 @@
 package yc.team8.baseball.post.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import yc.team8.baseball.member.domain.Member;
+import yc.team8.baseball.member.service.MemberService;
 import yc.team8.baseball.post.domain.Post;
 import yc.team8.baseball.post.dto.PostDto;
 import yc.team8.baseball.post.service.PostService;
@@ -17,10 +22,29 @@ public class PostController {
 
     @Autowired
     private PostService postService;
-
+    @Autowired
+    private MemberService memberService;
     // 각 구단의 게시판 창 띄우기
     @GetMapping("/{teamName}/{postType}")
-    public String showBoard(@PathVariable String teamName, @PathVariable String postType, Model model) {
+    public String showBoard(@PathVariable String teamName,
+                            @PathVariable String postType,
+                            HttpServletRequest request,
+                            RedirectAttributes rttr,
+                            Model model) {
+        // 잡담 게시판으로 접근하려 할 경우 -> 현재 유저가 자격이 있는지 확인
+        if(postType.equals("talk")) {
+            HttpSession session = request.getSession(false);
+            if(session == null)
+                return "redirect:/login";
+            Long id = (Long)session.getAttribute("memberID");
+            Member member = memberService.getMemberWithId(id);
+            if(!member.getTeam().equals(teamName)) {        // 자격이 없음
+                String msg = "You're not a fan of " + teamName + "!";
+                rttr.addFlashAttribute("msg", msg);
+                return "redirect:/" + teamName;
+            }
+        }
+
         model.addAttribute("teamName", teamName);
         model.addAttribute("postType", postType);
         // 해당 구단의 해당 게시판 글 목록 가져오기
@@ -53,10 +77,25 @@ public class PostController {
     }
     // 게시판에서 수정할 게시글 화면 띄우기
     @GetMapping("/{teamName}/{postType}/{id}/edit")
-    public String editPost(@PathVariable String teamName, @PathVariable String postType,
-                           @PathVariable Long id, Model model) {
+    public String editPost(@PathVariable String teamName,
+                           @PathVariable String postType,
+                           @PathVariable Long id,
+                           HttpServletRequest request,
+                           RedirectAttributes rttr,
+                           Model model) {
+        // 해당 게시글의 작성자가 접근하는 건지 검사
+        HttpSession session = request.getSession(false);
+        if(session == null)
+            return "redirect:/login";
+        Long member_id = (Long)session.getAttribute("memberID");
         // 수정할 데이터 가져오기
         Post target = postService.getPost(id);
+
+        if(target.getWriterId() != member_id) {     // 게시글 작성자가 아님
+            rttr.addFlashAttribute("msg", "Only writer can revise it.");
+            return "redirect:/" + teamName + "/" + postType + "/" + id;
+        }
+
         // 모델에 수정할 게시글, 구단 이름, 게시판 종류 저장하기
         model.addAttribute("post", target);
         model.addAttribute("teamName", teamName);
@@ -67,17 +106,32 @@ public class PostController {
     @PostMapping("/{teamName}/{postType}/{id}/update")
     public String updatePost(@PathVariable String teamName, @PathVariable String postType,
                              @PathVariable Long id, PostDto postDto) {
-        // 해당 게시물을 수정할 자격이 있는지 확인 필요 (글 작성자 확인)
-        //
-        //
-        Post saved = postService.updatePost(postDto);       // 게시물 업데이트 후 db에 저장된 게시물 리턴
+        Post saved = postService.updatePost(postDto);       // 게시물 업데이트 후 DB에 저장된 게시물 리턴
         return "redirect:/" + teamName + "/" + postType + "/" + saved.getId();       // 업데이트한 게시물 화면으로 리다이렉트
     }
     // 게시판에서 게시글 삭제
     @GetMapping("/{teamName}/{postType}/{id}/delete")
-    public String deleteInfo(@PathVariable String teamName, @PathVariable String postType,
-                             @PathVariable Long id) {
+    public String deleteInfo(@PathVariable String teamName,
+                             @PathVariable String postType,
+                             @PathVariable Long id,
+                             HttpServletRequest request,
+                             RedirectAttributes rttr) {
+        // 해당 게시글의 작성자가 접근하는 건지 검사
+        HttpSession session = request.getSession(false);
+        if(session == null)
+            return "redirect:/login";
+        Long member_id = (Long)session.getAttribute("memberID");
+        // 삭제를 원하는 데이터 가져오기
+        Post target = postService.getPost(id);
+
+        if(target.getWriterId() != member_id) {     // 게시글 작성자가 아님
+            rttr.addFlashAttribute("msg", "Only writer can delete it.");
+            return "redirect:/" + teamName + "/" + postType + "/" + id;
+        }
+
+        // 게시글 작성자 확인 후 게시글 삭제
         postService.deletePost(id);
+        rttr.addFlashAttribute("msg", "The deletion has been completed.");
         return "redirect:/" + teamName + "/" + postType;     // 해당 구단의 해당 게시판 목록으로 이동
     }
 }
